// dream-big-ui/src/app/Components/Star/types.ts
export interface Category {
    name: string;
    score: number;
    colour: string;
}


export interface PolygonPoints {
    spike: paper.Point;
    edgeL: paper.Point;
    edgeR: paper.Point;
    centre: paper.Point;


}
export interface Polygon {
    points: PolygonPoints

    highlight: boolean;
    category: Category;
    path: paper.Path;
}

export interface StarCoord {
    spike: paper.Point;
    edgeL: paper.Point;
    edgeR: paper.Point;
}